pyproject.toml
[project]
name = "trader-floor-ai-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Minimal runtime dependencies (aligned with requirements.txt)
    "gradio>=5.47.2,<6.0.0",
    "openai>=1.68.2,<2.0.0",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.0.1,<2.0.0",
    "pydantic>=2.9.0,<3.0.0",
    "requests>=2.32.3,<3.0.0",
    "pandas>=2.2.0,<3.0.0",
    "plotly>=5.24.0,<6.0.0",
    "polygon-api-client>=1.14.5,<2.0.0",
    "mcp>=1.15.0,<2.0.0",
    "mcp-server-fetch>=2025.1.17",
    "mcp[cli]>=1.5.0",
    # python-dateutil is a transitive dependency of pandas; keep implicit.
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

